pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "headmatch"
version = "0.1.0"
description = "PipeWire headphone measurement, EQ fitting, and headphone cloning toolkit"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "soundfile",
  "PyYAML",
]

[project.scripts]
headmatch = "headmatch.cli:main"
